+++ client/.env.example
@@ -5,4 +5,8 @@ REACT_APP_STORAGE_BUCKET =
 REACT_APP_MESSAGING_SENDER_ID =
 REACT_APP_APP_ID =
 
-FAST_REFRESH=false
+REACT_APP_EMAILJS_SERVICE_ID = 
+REACT_APP_EMAILJS_TEMPLATE_ID = 
+REACT_APP_EMAILJS_PUBLIC_KEY = 
+
+FAST_REFRESH = false
