--- client/src/assets/svg/Services/BrowserIcon.jsx
@@ -1,37 +1,0 @@
-import * as React from 'react';
-
-function SvgComponent(props) {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width={39.581}
-      height={39.58}
-      viewBox='0 0 39.581 39.58'
-      {...props}
-    >
-      <g data-name='Group 72'>
-        <path
-          data-name='Path 127'
-          d='M34.807 38.214H4.778a3.424 3.424 0 01-3.412-3.41V14.548a3.426 3.426 0 013.408-3.4h30.029a3.423 3.423 0 013.41 3.411v20.246a3.423 3.423 0 01-3.408 3.413z'
-          fill='#4cd5c5'
-        />
-        <g data-name='Group 69' fill='#0b093b'>
-          <path
-            data-name='Path 128'
-            d='M34.803 0H4.777A4.785 4.785 0 000 4.777v30.025a4.787 4.787 0 004.777 4.778h30.026a4.785 4.785 0 004.776-4.778V4.777A4.785 4.785 0 0034.803 0zm2.046 34.8a2.068 2.068 0 01-2.046 2.048H4.777A2.07 2.07 0 012.729 34.8V14.547a2.07 2.07 0 012.048-2.04h30.026a2.069 2.069 0 012.046 2.05zm0-24.552a4.775 4.775 0 00-2.046-.47H4.777a4.783 4.783 0 00-2.048.47V4.774a2.069 2.069 0 012.047-2.047h30.026a2.068 2.068 0 012.046 2.047z'
-          />
-          <path
-            data-name='Path 129'
-            d='M25.905 3.182a3.072 3.072 0 103.071 3.072 3.075 3.075 0 00-3.071-3.072zm0 4.378a1.307 1.307 0 111.307-1.305 1.314 1.314 0 01-1.305 1.306z'
-          />
-          <path
-            data-name='Path 130'
-            d='M32.673 3.182a3.072 3.072 0 103.07 3.072 3.073 3.073 0 00-3.07-3.072zm0 4.378a1.307 1.307 0 111.307-1.305 1.315 1.315 0 01-1.305 1.306z'
-          />
-        </g>
-      </g>
-    </svg>
-  );
-}
-
-export default SvgComponent;
