--- client/src/assets/svg/Services/RollerIcon.jsx
@@ -1,28 +1,0 @@
-import * as React from 'react';
-
-function SvgComponent(props) {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width={35.459}
-      height={43.42}
-      viewBox='0 0 35.459 43.42'
-      {...props}
-    >
-      <g data-name='Group 70'>
-        <path
-          data-name='Path 121'
-          d='M24.571 12.115H4.778a3.425 3.425 0 01-3.412-3.411V4.78a3.425 3.425 0 013.411-3.415h19.792a3.427 3.427 0 013.413 3.412v3.924a3.424 3.424 0 01-3.408 3.414z'
-          fill='#f2b300'
-        />
-        <path
-          data-name='Path 122'
-          d='M30.981 4.552h-1.647a4.782 4.782 0 00-4.765-4.55H4.779A4.787 4.787 0 00.001 4.781v3.922a4.787 4.787 0 004.778 4.778h19.79a4.785 4.785 0 004.776-4.778V7.282h1.636a1.836 1.836 0 011.748 1.876 4.152 4.152 0 01-2.31 3.309l-13.158 5.54a6.889 6.889 0 00-3.98 5.995v1.5a4.784 4.784 0 00-3.555 4.612v8.531a4.782 4.782 0 004.778 4.775h.339a4.784 4.784 0 004.776-4.776v-8.531a4.782 4.782 0 00-3.611-4.626V24a4.292 4.292 0 012.311-3.478l13.16-5.542a6.75 6.75 0 003.98-5.824 4.553 4.553 0 00-4.478-4.604zm-4.364 4.15a2.069 2.069 0 01-2.048 2.047H4.779a2.069 2.069 0 01-2.048-2.047V4.78a2.07 2.07 0 012.048-2.049h19.79a2.07 2.07 0 012.048 2.049v3.922zM16.89 30.114v8.531a2.067 2.067 0 01-2.047 2.046h-.339a2.07 2.07 0 01-2.05-2.046v-8.531a2.071 2.071 0 012.05-2.048h.339a2.066 2.066 0 012.047 2.047z'
-          fill='#0b093b'
-        />
-      </g>
-    </svg>
-  );
-}
-
-export default SvgComponent;
