+++ client/src/config/email.ts
@@ -1,0 +1,13 @@
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+const EMAILJS_SERVICE_ID = process.env.REACT_APP_EMAILJS_SERVICE_ID || '';
+const EMAILJS_TEMPLATE_ID = process.env.REACT_APP_EMAILJS_TEMPLATE_ID || '';
+const EMAILJS_PUBLIC_KEY = process.env.REACT_APP_EMAILJS_PUBLIC_KEY || '';
+
+export const emailJSConfig = {
+  serviceID: EMAILJS_SERVICE_ID,
+  templateID: EMAILJS_TEMPLATE_ID,
+  publicKey: EMAILJS_PUBLIC_KEY,
+};
