+++ client/src/containers/LandingPage.tsx
@@ -3,7 +3,6 @@ import { Box } from '@mui/material';
 
 import Contact from '../components/landing-page/sections/Contact';
 import Header from '../components/landing-page/sections/Header';
-import Pricing from '../components/landing-page/sections/Pricing';
 import Services from '../components/landing-page/sections/Services';
 import Footer from '../components/landing-page/sections/Footer';
 import FadeInSection from '../components/utils/FadeInSection';
@@ -15,9 +14,6 @@ const LandingPage = () => (
     </FadeInSection>
     <FadeInSection>
       <Services />
-    </FadeInSection>
-    <FadeInSection>
-      <Pricing />
     </FadeInSection>
     <FadeInSection>
       <Contact />
